Examples import from Theano now.
Also ranamed non-importable files with '-' in the name.
