+++ examples/singleneuron.py
@@ -1,4 +1,4 @@
-import nef
+import nef.nef_theano as nef
 
 net=nef.Network('Single Neuron')      # Create the network
